feat: separate cli and defaults from base functionality
The cli code now lives in cli.js, defaults in defaults.json and base
functionality in index.js. As a result, standard-version can now be
used as a library. See cli.js for usage. Calls to exec and
handleExecError have been refactored into a single handledExec
function that takes an error and success callback separately.