docs(FAQ): explain why is chore entries not in my CHANGELOG
I was surprised to see people asking for this since it made sense to me when I looked at it the first time. But too many people asking for it so maybe worth to add this? Mostly I copied @nexdrew's comment in the issue https://github.com/conventional-changelog/standard-version/issues/135.

Closes https://github.com/conventional-changelog/standard-version/issues/135