fix(cli): Ensure that the skip and scripts cli arguments are first parsed before they are evaluated
Without first JSON parsing the arguments of --scripts and --skip the typeof check within the
command.js was always throwing an error. It seems like yargs doesn't perform the parse for you, so
I've added the step to parse (and check for valid json). Now I'm able to do the commands that #373
is highlighting.

#373